Update to _Cell._validate
Corrected cell height/width validation method to evaluate dimensions of cells and cell text based on pixels.  Previous implementation could result in erroneous measurements in validation, requiring correction to allow for finer granularity in units along with additional tolerance for variation in pixel calculation (±3 pixels) to prevent unintended error.  The calculated maximum dimensions for cells are the ceiling value to the nearest pixel, cell padding is the floor value of the combined padding per side to the nearest pixel, and text width is the floor width to the nearest pixel.  At 300 DPI, rendered text may now be up to approximately 3 pixels wider (0.01 in.) to accommodate previous error scenario.